推荐接口

+++ src/main/java/com/hqq/colorful_life/model/service/serviceImpl/RecommendationServiceImpl.java
@@ -1,10 +1,19 @@
 package com.hqq.colorful_life.model.service.serviceImpl;
 
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
+import com.hqq.colorful_life.exception.ExceptionEnum;
+import com.hqq.colorful_life.exception.UniteException;
+import com.hqq.colorful_life.model.dao.CreateItemMapper;
+import com.hqq.colorful_life.model.domain.CreateItem;
 import org.springframework.stereotype.Service;
 import javax.annotation.Resource;
 import com.hqq.colorful_life.model.domain.Recommendation;
 import com.hqq.colorful_life.model.dao.RecommendationMapper;
 import com.hqq.colorful_life.model.service.RecommendationService;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * @author Qingqing.He
@@ -16,13 +25,26 @@ public class RecommendationServiceImpl implements RecommendationService {
     @Resource
     private RecommendationMapper recommendationMapper;
 
+    @Resource
+    CreateItemMapper createItemMapper;
+
     @Override
     public int deleteByPrimaryKey(Integer id) {
+        if (id == null) {
+            throw new UniteException(ExceptionEnum.PARAM_NOT_NULL);
+        }
         return recommendationMapper.deleteByPrimaryKey(id);
     }
 
     @Override
     public int insert(Recommendation record) {
+        if (record == null) {
+            throw new UniteException(ExceptionEnum.PARAM_NOT_NULL);
+        }
+        CreateItem createItem = createItemMapper.selectByPrimaryKey(record.getCreateId());
+        if (createItem == null) {
+            throw new UniteException(ExceptionEnum.REQUEST_PARAM_ERROR);
+        }
         return recommendationMapper.insert(record);
     }
 
@@ -44,6 +66,22 @@ public class RecommendationServiceImpl implements RecommendationService {
     @Override
     public int updateByPrimaryKey(Recommendation record) {
         return recommendationMapper.updateByPrimaryKey(record);
+    }
+
+    @Override
+    public PageInfo selectListAll(Integer pageNum, Integer pageSize) {
+
+        PageHelper.startPage(pageNum,pageSize);
+        List<Recommendation> recommendations = recommendationMapper.selectListAll();
+        ArrayList<CreateItem> createItems = new ArrayList<>();
+        for (int i = 0; i < recommendations.size(); i++) {
+            Recommendation recommendation = recommendations.get(i);
+            CreateItem createItem = createItemMapper.selectByPrimaryKey(recommendation.getCreateId());
+            createItems.add(createItem);
+        }
+        PageInfo<CreateItem> createItemPageInfo = new PageInfo<>(createItems);
+        return createItemPageInfo;
+
     }
 
 }
